% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acessoInternet.R
\name{acessoInternet}
\alias{acessoInternet}
\title{Demografico: obtem dados dos participantes sobre acesso a Internet}
\usage{
acessoInternet(inscritos)
}
\arguments{
\item{inscritos}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com quantidade de respostas por opcao
}
\description{
Demografico: obtem dados dos participantes sobre acesso a Internet
}
\author{
Ivan L M Ricarte
}

