% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acessoPortal.R
\name{acessoPortal}
\alias{acessoPortal}
\title{Demografico: frequencia de acesso ao portal Saude baseada em evidencias}
\usage{
acessoPortal(inscritos)
}
\arguments{
\item{inscritos}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com quantidade de respostas por opcao
}
\description{
Demografico: frequencia de acesso ao portal Saude baseada em evidencias
}
\author{
Ivan L M Ricarte
}

