% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/avaliaBeneficio.R
\name{avaliaBeneficio}
\alias{avaliaBeneficio}
\title{IAM: Qual beneficio esperado pelo uso da evidencia (questao 4)}
\usage{
avaliaBeneficio(dados)
}
\arguments{
\item{dados}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com proporcao de respostas por opcao
}
\description{
IAM: Qual beneficio esperado pelo uso da evidencia (questao 4)
}
\author{
Ivan L M Ricarte
}

