% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/avaliaImpacto.R
\name{avaliaImpacto}
\alias{avaliaImpacto}
\title{IAM: Qual impacto cognitivo da evidencia (questao 1)}
\usage{
avaliaImpacto(dados)
}
\arguments{
\item{dados}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com proporcao de respostas por opcao
}
\description{
IAM: Qual impacto cognitivo da evidencia (questao 1)
}
\author{
Ivan L M Ricarte
}

