% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/avaliaUso.R
\name{avaliaUso}
\alias{avaliaUso}
\title{IAM: Qual uso potencial da evidencia (questao 3)}
\usage{
avaliaUso(dados)
}
\arguments{
\item{dados}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com proporcao de respostas por opcao
}
\description{
IAM: Qual uso potencial da evidencia (questao 3)
}
\author{
Ivan L M Ricarte
}

