% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/byProfession.R
\name{byProfession}
\alias{byProfession}
\title{Tabulate by profession}
\usage{
byProfession(profdata)
}
\arguments{
\item{profdata}{data frame with professional data}
}
\value{
data frame with profession and number of professionals
(total and by level of attention)
}
\description{
Tabulate by profession
}
\author{
Ivan L M Ricarte
}

