% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculaCrii.R
\name{calculaCrii}
\alias{calculaCrii}
\title{IAM: Qual relevancia clinica da evidencia (questao 2)}
\usage{
calculaCrii(dados)
}
\arguments{
\item{dados}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com proporcao de respostas por opcao e indice de relevancia
}
\description{
IAM: Qual relevancia clinica da evidencia (questao 2)
}
\author{
Ivan L M Ricarte
}

