% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extrairBases.R
\name{extrairBases}
\alias{extrairBases}
\title{Resumos: obtem bases de evidencias utilizadas}
\usage{
extrairBases(resumos)
}
\arguments{
\item{resumos}{data frame com dados importados do instrumento Resumos}
}
\value{
data frame com nome da base e os resumos extraidos dessa base
}
\description{
Resumos: obtem bases de evidencias utilizadas
}
\author{
Ivan L M Ricarte
}

