% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extrairComentarios.R
\name{extrairComentarios}
\alias{extrairComentarios}
\title{IAM: Comentarios sobre a evidencia (campo aberto)}
\usage{
extrairComentarios(dados)
}
\arguments{
\item{dados}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com todos os comentarios por resumo
}
\description{
Efeito colateral: criacao de arquivos de texto com comentarios para cada evidencia
}
\author{
Ivan L M Ricarte
}

