% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/participantAssessmentList.R
\name{formataCPF}
\alias{formataCPF}
\title{Ajusta formatacao de CPF}
\usage{
formataCPF(cpfs)
}
\arguments{
\item{cpfs}{vetor de CPF com grafia nao uniforme}
}
\value{
vetor de CPF com formato ajustado
}
\description{
Ajusta formatacao de CPF
}
\author{
Ivan L M Ricarte
}

