% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/formatName.R
\name{formatName}
\alias{formatName}
\title{Capitalize words in proper names, lowercase connectives}
\usage{
formatName(nomes)
}
\arguments{
\item{nomes}{names, as inserted by users}
}
\value{
formatted names
}
\description{
Capitalize words in proper names, lowercase connectives
}
\author{
Ivan L M Ricarte
}

