% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRedcap.R
\name{getRedcap}
\alias{getRedcap}
\title{Get data directly from REDCap}
\usage{
getRedcap(url = "http://redcap.fmrp.usp.br/api/", mytoken)
}
\arguments{
\item{url}{URI for REDCap API interface}

\item{myToken}{token for REDCap project}
}
\value{
data frame full data set
}
\description{
Get data directly from REDCap
}
\author{
Ivan L M Ricarte
}

