% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRedcapFromFile.R
\name{getRedcapFromFile}
\alias{getRedcapFromFile}
\title{Get data from file exported by REDCap}
\usage{
getRedcapFromFile(filename)
}
\arguments{
\item{filename}{name of csv file with REDCap data}
}
\value{
data frame com todos os dados importados
}
\description{
Get data from file exported by REDCap
}
\author{
Ivan L M Ricarte
}

