% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/participantAssessmentList.R
\name{participantAssessmentList}
\alias{participantAssessmentList}
\title{Demographic + IAM: get all participant names and quantity of assessments for each}
\usage{
participantAssessmentList(fullDataset, drop = TRUE)
}
\arguments{
\item{drop}{logic, TRUE if output list excludes participants with no assessment}

\item{data}{data frame with full REDCap exported data}
}
\value{
data frame name, document (CPF) number, CPF validity, and number of assessments
}
\description{
This function should be used to obtain the data that will be on certificates
}
\author{
Ivan L M Ricarte
}

