% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/personalData.R
\name{personalData}
\alias{personalData}
\title{Extract personal data (age, sex) from full data set}
\usage{
personalData(fullDataset, referenceDate = NA)
}
\arguments{
\item{fullDataset}{data frame with REDCap exported data}

\item{referenceDate}{date (YYYY-MM-DD) to compute age; if NA, assumes today}
}
\value{
data frame with id, age e sex for each participant.
}
\description{
Extract personal data (age, sex) from full data set
}
\author{
Ivan L M Ricarte
}

