% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/professionalData.R
\name{professionalData}
\alias{professionalData}
\title{Demographic: participants' professional experience}
\usage{
professionalData(fullDataset)
}
\arguments{
\item{fullDataset}{data frame with REDCap exported data}
}
\value{
data frame with participant id,
}
\description{
Demographic: participants' professional experience
}
\author{
Ivan L M Ricarte
}

