% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recursosInformacionais.R
\name{recursosInformacionais}
\alias{recursosInformacionais}
\title{Demografico: obtem dados dos participantes sobre recursos informacionais usados}
\usage{
recursosInformacionais(inscritos)
}
\arguments{
\item{inscritos}{data frame com todos os dados dos participantes inscritos}
}
\value{
data frame com quantidade de respostas por opcao
}
\description{
Demografico: obtem dados dos participantes sobre recursos informacionais usados
}
\author{
Ivan L M Ricarte
}

