% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/participantAssessmentList.R
\name{verificaCPF}
\alias{verificaCPF}
\title{Verifica dígitos de um CPF}
\usage{
verificaCPF(cpf)
}
\arguments{
\item{cpf}{um CPF no formato ###.###.###-##}
}
\value{
TRUE se digitos verificadores estao coerentes com CPF
}
\description{
Verifica dígitos de um CPF
}
\author{
Ivan L M Ricarte
}

